Add files via upload
Smart City Management
Smart City is an application in Java programming language to store the details of a city. The main purpose of this project to help the tourists and other visitors to the city by providing information about hotels, transportation facilities, air ticket booking, shopping details, city news, etc. The Smart City Java Project serves the visitors as a guide of the city.

Technologies used
Used Java programming language, JavaFX for GUI and Mysql.